requirements.txt
pandas==2.2.2
python-dotenv==1.0.1
requests==2.32.3
google-api-python-client==2.137.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.32
boto3==1.34.158
